setup/teardown DiscoverAndQueue data
